+++ app/Http/Controllers/LinkController.php
@@ -8,6 +8,7 @@ use App\Models\Link;
 use Illuminate\Support\Facades\Redirect;
 use App\Http\Requests\StoreLinkPostRequest;
 use Auth;
+use Illuminate\Support\Facades\Session;
 
 class LinkController extends Controller
 {
@@ -18,14 +19,16 @@ class LinkController extends Controller
      * @return mixed
      */
     public function create(Request $request) {
+        if ( !$this->checkHTML($request['html_id']) ) {
+            Session::flash('success', trans('errors.LS40401_UNKNOWN'));
+            return Redirect::to('/');
+        }
         if ( $this->userDisable(Auth::user()->status, 'edit') ) {
-            $create_data = ['title' => $request['title'], 'url' => $request['url'], 'status' => $request['status'], 'description' => $request['description']];
+            $create_data = ['title' => $request['title'], 'url' => $request['url'], 'status' => $request['status'], 'description' => $request['description'], 'thumb' => '/storage/uploads/default.png'];
             if(!empty($_FILES['thumb']['name'])){
                 $img_res = $this->uploadImage($this->base_path);
                 if($img_res['success'] == 1) {
                     $create_data['thumb'] = $img_res['result'];
-                }else{
-                    $create_data['thumb'] = '/storage/uploads/default.png';
                 }
             }
             $result = Link::create($create_data);
@@ -47,6 +50,10 @@ class LinkController extends Controller
      * @return mixed
      */
     public function edit(Request $request) {
+        if ( !$this->checkHTML($request['html_id']) ) {
+            Session::flash('success', trans('errors.LS40401_UNKNOWN'));
+            return Redirect::to('/');
+        }
         if ( $this->userDisable(Auth::user()->status, 'edit') ) {
             $update_data = ['title' => $request['title'], 'url' => $request['url'], 'status' => $request['status'], 'description' => $request['description']];
             if(!empty($_FILES['thumb']['name'])){
@@ -74,6 +81,10 @@ class LinkController extends Controller
      * @return string
      */
     public function removed(Request $request) {
+        if ( !$this->checkHTML($request['html_id']) ) {
+            $record = ['success' => 0, 'result' => trans('errors.LS40401_UNKNOWN')];
+            return json_encode($record);
+        }
         if ( $this->userDisable(Auth::user()->status, 'remove') ) {
             $result = Link::destroy((int)$request['id']);
             if ( $result ) {
@@ -85,5 +96,16 @@ class LinkController extends Controller
             $record = ['success' => 0, 'result' => trans('validation.user.disabled_power'),['op' => 'Remove']];
         }
         return json_encode($record);
+    }
+    /**
+     * Ensure can through this function.
+     *
+     * @param $html
+     * @return bool
+     */
+    private function checkHTML($html) {
+        if(!empty($html) && $html == 'link')
+            return true;
+        return false;
     }
 }
